Make Glyph requests from the Map Thread
Now the requests are bound to the Map Thread loop and
we can cancel it quickly when shutting down or changing
style.

Previously requests were made by the workers and we would
have to wait them to finish because workers are destroyed
before the GlyphStore. This was giving the impression that
the application was frozen (in fact it was, momentarily).